requirements.txt
# Web Framework
flask>=2.0.1
flask-cors>=3.0.10
gunicorn>=20.1.0

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pandas>=1.3.0
openpyxl>=3.0.7
Pillow>=8.3.1
pytesseract>=0.3.8

# Knowledge Graph and Vector DB
networkx>=2.6.3
rdflib>=6.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning and API
transformers>=4.30.0
torch>=2.0.0
requests>=2.26.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
tqdm>=4.65.0
click>=8.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=22.3.0
isort>=5.10.1
flake8>=4.0.1
mypy>=0.950